Added GameClient and GameServer
When pressing play in the lobby the game now creates a Client scene and
also a Server scene if the current client is the lobby owner. Then the
transforms of some cubes are synchronized from the server for all
clients. It is proof that unity can run two scenes at the same time
where one of them is the server.