Networking for transforms in the hierarchy works
ServerObjects have to be stored in the resources folder and will be
automatically synchronized with all clients in the hierarchy. Parenting
and destruction of these objects will be synchronized too. Basically all
the objects are now identified by the server side transform instance id
of a ServerObject. The messages are sent over different channels - each
representing the type of the message. JSON is used for the serialization
of these messages.