Fixed NullReferenceException on client only version
Also it is important that all objects in the Client scene are on the
Layer "Client" and all objects in the Server scene are on the Layer
"Server". Otherwise the cameras can see both (the client and the server)
objects at the same time. Right now Display1 shows the Client side and
Display2 shows the Server side.