Updated lobby & demo, added MessageEvents for the Server
The ClientManager now has two messageEvents - one for the client and one
for the server. Messages for the server are now sent on different
channels than the ones for the client. This way there is no need for a
new NetworkMessageType for each direction. In the lobby now only the
lobby owner checks if everyone is ready and then sends a message to all
members to start the game. By pressing Space in the demo scene, the
client will tell the server to push all rigidbodies up.