Added NetworkBehaviour class for easy networking
The basic networking workflow is like this right now: Create a new
object with a ServerObject script on it, add a new NetworkMessageType
for your communication and create a new script that derives from
NetworkBehaviour. Then all the Client side and Server side management
can be done in this one script. There are seperate functions that should
be overwritten for the Client and the Server. Right now these are
StartClient, StartServer, UpdateClient, UpdateServer, OnDestroyClient,
OnDestroyServer. These functions replace the Start, Update and OnDestroy
functions of the MonoBehaviour. For custom functions you should check
the onServer attribute. If you want to override e.g. the Start function
you have to call base.Start() inside.