Added template for network behaviour
Just right click in a folder in the inspector and select Create/C#
Script NetworkBehaviour. I also improved the StartClient and StartServer
methods and added the names of the scenes to load in the lobby as a
public string on the LobbyManager. Also the 32bit version of the
Facepunch.Steamworks SDK is now disabled since it likely is the reason
why the Editor crashes sometimes.