Added VoiceChat NetworkBehaviour (not functional yet)
Also replaced Vector3 and Quaternion inside the MessageServerObject struct by floats because Unity does not allow Serialization for these types. Decided not to replace the ByteSerializer class by the C# BinaryFormatter because this formatter adds a lot of overhead that impacts the amount of data we can send a lot (e.g instead of 14 server objects in one UDP message we could only send 2).